requirements.txt
aniso8601==10.0.0
argcomplete==3.5.3
asgiref==3.8.1
Beaker==1.12.1
beautifulsoup4==4.12.3
blinker==1.9.0
blivet==3.11.0
blivet-gui==2.6.0
build==1.2.2.post1
CacheControl==0.14.1
certifi==2025.1.31
cffi==1.17.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
cloudinary==1.42.2
colorama==0.4.6
contourpy==1.3.1
crashtest==0.4.1
crypt_r==3.13.1
cryptography==43.0.0
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.3.2
distlib==0.3.9
distro==1.9.0
Django==5.1.6
djangorestframework==3.15.2
dnf==4.22.0
dnspython==2.7.0
dulwich==0.21.7
fastjsonschema==2.21.1
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.16.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
fros==1.1
gi-docgen==2024.1
greenlet==3.1.1
humanize==4.9.0
idna==3.10
imageio==2.37.0
installer==0.7.0
iso639==0.1.4
itsdangerous==2.2.0
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.5
keyring==24.3.1
kiwisolver==1.4.8
langtable==0.0.68
lazy_loader==0.4
libcomps==0.1.21
libdnf==0.73.4
lxml==5.2.1
Mako==1.2.3
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
meson==1.5.1
more-itertools==10.5.0
msgpack==1.1.0
networkx==3.4.2
nftables==0.1
numpy==2.2.3
olefile==0.47
opencv-python==4.11.0.86
packaging==24.1
pandas==2.2.3
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
pkginfo==1.12.0
platformdirs==4.3.6
ply==3.11
podman-compose==1.3.0
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
productmd==1.43
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pyenchant==3.2.2
Pygments==2.18.0
PyGObject==3.48.2
pyinotify==0.9.6
pykickstart==3.58
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyparted==3.13.0
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
pytz==2025.1
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
RapidFuzz==3.11.0
regex==2024.9.11
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
requests-toolbelt==1.0.0
rpm==4.20.0
scikit-image==0.25.1
scipy==1.15.1
SecretStorage==3.3.3
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.2.0
shellingham==1.5.4
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
smartypants==2.0.1
sos==4.8.2
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
systemd-python==235
Tempita==0.5.2
tifffile==2025.1.10
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
typogrify==2.0.7
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.28.0
Werkzeug==3.1.3
xkbregistry==0.3
